requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
https://github.com/allanwright/media-classifier-core/archive/v1.2.2.tar.gz
graphene
spacy==2.3.5
azure-storage-queue
titlecase